new abstrac class binding control => moved some fields from namedValueTemplate
